requirements.txt
numpy
pandas
scipy
sklearn
pytorch

